% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatCompR.R
\name{benchmarks}
\alias{benchmarks}
\title{Benchmark R and Rcpp functions.}
\description{
the useage of C functions \code{gibbsC} .
}
\examples{
\dontrun{
data(data)
attach(data)
gibbC=gibbsC(200,1)
ts <- microbenchmark(gibbR=gibbsR(200,1),
gibbC=gibbsC(200,1))
summary(ts)[,c(1,3,5,6)]
}
}
