% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatCompR.R
\name{positive.mean}
\alias{positive.mean}
\title{Calculate positive mean  using R.}
\usage{
positive.mean(X, d)
}
\arguments{
\item{X}{the observed data (numeric)}

\item{d}{the deductible (numeric)}
}
\value{
the mean of X-d 's positive part\code{n}
}
\description{
the mean of X-d 's positive part.
}
\examples{
\dontrun{
set.seed(123)
X<-rnorm(100)
positive.mean(X,3)
}
}
