% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatCompR.R
\name{RM}
\alias{RM}
\title{Risk measure.}
\usage{
RM(X, alpha, measure = c("VaR", "CVaR", "Distortion", "Shortfall"), lambda)
}
\arguments{
\item{X}{the Observed data (numeric)}

\item{alpha}{the confidence level (numeric)}

\item{measure}{the type of risk measure with 'VaR','CVaR','Distortion','Shortfall'  (logical)}

\item{lambda}{the risky parameter(numeric)}
}
\value{
a random sample of size \code{n}
}
\description{
The function can be used to calculate the risk measure such as 'VaR','CVaR','Distortion','Shortfall'.
}
\examples{
\dontrun{
set.seed(1234)
X<-rnorm(100,0,4)
RM(X,0.9,measure='VaR')
RM(X,0.9,measure='CVaR')
RM(X,measure='Distortion')
RM(X,measure='Shortfall',lambda=20)
}
}
