\name{StatComp22001-package}
\alias{StatComp22001-package}
\docType{package}
\title{
This package provides some methods to calculate Risk Measure
}
\description{
Several R and Rcpp function are included in this package for illustration.You can use it to calculate some risk measure such as VaR,CVaR,distortion,shortfall risk measure. \code{RM},and Rcpp function Gibbs sampler  \code{gibbsC}.
}
\author{
ll22001
Maintainer: ll22001 <459435706@qq.com> }
\keyword{ package }
